+++ psutil/_pswindows.py
@@ -410,7 +410,7 @@ def boot_time():
     # value which may have a 1 second fluctuation, see:
     # https://github.com/giampaolo/psutil/issues/1007
     global _last_btime
-    ret = cext.boot_time()
+    ret = float(cext.boot_time())
     if abs(ret - _last_btime) <= 1:
         return _last_btime
     else: